+++ vscode/CHANGELOG.md
@@ -9,6 +9,7 @@ Starting from `0.2.0`, Cody is using `major.EVEN_NUMBER.patch` for release versi
 ### Added
 
 - Chat: A new chat model selection dropdown that allows selecting between different chat models when connected to the sourcegraph.com instance. [pull/1676](https://github.com/sourcegraph/cody/pull/1676)
+- Chat: New button in editor title for restarting chat session in current chat panel. [pull/1687](https://github.com/sourcegraph/cody/pull/1687)
 
 ### Fixed